changed AuthFilter

+++ src/main/java/controller/DeleteProductServlet.java
@@ -1,4 +1,4 @@
-package controller.servlet;
+package controller;
 
 import factory.ProductServiceFactory;
 import org.apache.log4j.Logger;
@@ -9,6 +9,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 @WebServlet(value = "/products/delete")
@@ -20,13 +21,18 @@ public class DeleteProductServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request,
                          HttpServletResponse response) throws ServletException, IOException {
-        String id = request.getParameter("id");
-        if (id != null) {
-            productService.deleteProduct(Long.valueOf(id));
-            logger.info("Product { id = " + id + "} is deleted in db.");
+        final HttpSession session = request.getSession();
+        final String role = (String) session.getAttribute("role");
+        if (role.equals("admin")) {
+            String id = request.getParameter("id");
+            if (id != null) {
+                productService.deleteProduct(Long.valueOf(id));
+                logger.info("Product { id = " + id + "} is deleted in db.");
+            }
+            response.sendRedirect("/products");
+        } else {
+            response.sendRedirect("/");
         }
-        response.sendRedirect("/products");
-        response.setStatus(HttpServletResponse.SC_OK);
     }
 
 }
