+++ src/main/java/controller/DeleteUserServlet.java
@@ -1,4 +1,4 @@
-package controller.servlet;
+package controller;
 
 import factory.UserServiceFactory;
 import org.apache.log4j.Logger;
@@ -9,9 +9,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-@WebServlet(value = "/users/delete")
+@WebServlet(value = "/admin/users/delete")
 public class DeleteUserServlet extends HttpServlet {
 
     private static final UserService userService = UserServiceFactory.getUserService();
@@ -21,12 +22,17 @@ public class DeleteUserServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request,
                          HttpServletResponse response) throws ServletException, IOException {
         String id = request.getParameter("id");
-        if (id != null) {
-            userService.deleteUser(Long.valueOf(id));
-            logger.info("User { id = " + id + "} is deleted in db.");
+        final HttpSession session = request.getSession();
+        final String role = (String) session.getAttribute("role");
+        if (role.equals("admin")) {
+            if (id != null) {
+                userService.deleteUser(Long.valueOf(id));
+                logger.info("User { id = " + id + "} is deleted in db.");
+            }
+            response.sendRedirect("/admin/users");
+        } else {
+            response.sendRedirect("/");
         }
-        response.sendRedirect("/users");
-        response.setStatus(HttpServletResponse.SC_OK);
     }
 
 }
